added crappy bilinear filter

+++ src/player/rays_alt.c
@@ -30,12 +30,12 @@ void	calculate_ray_stuff(t_ray *ray, t_player *player, double gradient, double c
 	if (face_mod == 3 || face_mod == 0)
 	{
 		ray->intcpt = get_horizontal_int(ray->intcpt.y, gradient, c);
-		ray->pos = (int)(fmod(ray->intcpt.x, 1) * ray->texture->x);
+		ray->pos = (fmod(ray->intcpt.x, 1) * ray->texture->x);
 	}
 	else
 	{
 		ray->intcpt = get_vertical_int(ray->intcpt.x, gradient, c);
-		ray->pos = (int)(fmod(ray->intcpt.y, 1) * ray->texture->x);
+		ray->pos = (fmod(ray->intcpt.y, 1) * ray->texture->x);
 	}
 	ray->distance = get_cam_distance(player->pos, player->angle + M_PI_2, ray->intcpt);
 	// ray->distance = vector_distance(player->pos, ray->intcpt);
